backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
playwright
httpx
daytona
certifi
supabase
